LPS-127650 Merge useBackgroundImageQueries with loadBackgroundImage
